+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## [1.1.0-dev.3](https://github.com/mvaishak/letterboxd-morphe-patches/compare/v1.1.0-dev.2...v1.1.0-dev.3) (2026-09-02)
+
+### 🐛 Bug Fixes
+
+* Remove the template example patch ([3f965e5](https://github.com/mvaishak/letterboxd-morphe-patches/commit/3f965e55a306ce20162030bdd2f0ecfad3de8a26))
+
+## [1.1.0-dev.2](https://github.com/mvaishak/letterboxd-morphe-patches/compare/v1.1.0-dev.1...v1.1.0-dev.2) (2026-09-02)
+
+### 🐛 Bug Fixes
+
+* Flatten Material You chrome and fix sheet/divider tones ([719d6f7](https://github.com/mvaishak/letterboxd-morphe-patches/commit/719d6f7e73c0fee7789f7946e6c0d30d83c4ecd3))
+
+## [1.1.0-dev.1](https://github.com/mvaishak/letterboxd-morphe-patches/compare/v1.0.0...v1.1.0-dev.1) (2026-09-02)
+
+### ✨ New Features
+
+* Add Material You theme patch for Letterboxd ([8e52fc3](https://github.com/mvaishak/letterboxd-morphe-patches/commit/8e52fc3052dee51798192eae3fd2d080a177a7c4))
+
 ## 1.0.0 (2026-09-02)
 
 ### ✨ New Features